Move parser methods to ParserUtil.java
